feat: add new options CLI entrypoint definition.
